Complete Phase 7A.1: Security Critical Fixes - Encryption Key Management
✅ PHASE 7A.1 COMPLETE - All security critical fixes implemented and tested

## Security Improvements Implemented:
- ✅ Secure encryption key failure modes (no plaintext fallback)
- ✅ Production environment key export restrictions
- ✅ Comprehensive environment detection (dev vs prod)
- ✅ 10 comprehensive security tests passing
- ✅ Proper SecurityError inheritance and handling

## Web Demo Improvements:
- ✅ Simplified startup script (15 lines vs 200+ lines)
- ✅ Comprehensive debug/troubleshooting script
- ✅ Automatic port cleanup and conflict resolution
- ✅ Backend starts in ~2 seconds (vs 30+ second timeouts)
- ✅ Health check validation and API testing

## Regression Testing Results:
- ✅ 368/368 tests passing (100% success rate)
- ✅ All core demos working perfectly
- ✅ All examples working correctly
- ✅ Web demo fully functional
- ✅ No regressions in any functionality

## Files Modified:
- src/stinger/core/api_key_manager.py: Security improvements
- demos/web_demo/start_demo.py: Simplified startup
- demos/web_demo/debug_startup.py: QA troubleshooting tool
- tests/test_api_key_manager_security.py: 10 security tests
- Various web demo improvements and fixes

## Ready for Phase 7A.2: Architecture Cleanup
