+++ .cursor
@@ -174,6 +174,15 @@ GITHUB_REPO=virtualsteve-star/stinger
 # Run 'date' command to get the current date/time before writing dates in documentation
 # Never assume or guess dates - always verify with the system date
 
+# PROCESS TESTING RULES
+# - **NEVER use `command &` with Bash tool** - causes hanging when trying to capture output
+# - **For long-running services:** test endpoints directly, don't background processes
+# - **If testing startup:** use wrapper scripts that exit quickly or test specific functionality
+# - **Examples:**
+#   - ❌ `python3 server.py &` (will hang)
+#   - ✅ `curl http://localhost:8000/health` (test endpoint)
+#   - ✅ Start process, let user interrupt, then verify it worked
+
 # ---
 # API KEY ENVIRONMENT VARIABLE HANDLING (Stinger)
 #
